+++ src/net/tmwa/tradehandler.cpp
@@ -109,7 +109,7 @@ void TradeHandler::respond(bool accept)
         PlayerInfo::setTrading(false);
 
     MessageOut outMsg(CMSG_TRADE_RESPONSE);
-    outMsg.writeInt8(accept ? 3 : 4);
+    outMsg.writeInt8(static_cast<int8_t>(accept ? 3 : 4));
 }
 
 void TradeHandler::addItem(Item *item, int amount)
